// src/models/CryptoDataManagement.ts
export interface CoinMarketsResponse {
    id: string;
    name: string;
    symbol: string;
    current_price: number;
    market_cap: number;
    total_volume: number;
    price_change_percentage_24h?: number;
    circulating_supply?: number;
    total_supply?: number;
    last_updated?: string;
}